Add action and event classes
Implement the card module redesign (which uses a raw and
expressive event API rather than a high-level convenience
API) by:

 * Removing the old, external state based system (i.e., all files
   under gradian/engine/).

 * Creating the event classes that a module can process, and the
   action classes a client can perform.

NOTE: an additional design decision was made to omit native
face-down card functionality and, instead, have it completely
controlled by the game module. This choice was made to simplify
the logic and give more control to the game module.