Add module loader and hookup websockets to manager
Add a module loader which searches for all the modules (zip
files with an `entry.py` module within them) within `modules/`
and loads them -- invalid modules are ignored.

Add the websocket server, which listens for new websocket
connections, and forward them to the appropriate `EngineManager`
service:

 * For `/create/:module_id` websockets, forward them to the
   create function.

 * For `/join/:game_id` websockets, forward them to the join
   function.

 * Drop all other websockets.

NOTE: no modules are included as part of this commit and, as a
result, it is unrunnable in its current state.