Add basic engine unit tests
Create a mock WebSocket connection class to perform some basic
unit tests on the engine. The mock connection is not perfect,
and will not replicate the behaviour of a WebSocket perfectly,
but should be to a good enough degree that the engine should
work the same for real WebSocket connections.

Additionally, make some miscellaneous bug fixes and consistency
improvements (e.g., use `coll` in the WebSocket messages rather
than have a mix of `coll` and `collection`).