Create examplar Rummy card module
Create an exemplar card module implementation, which implements
the card game Rummy -- this game was chosen due to its simplicity.

Additional changes were also made to facilitate this (e.g.,
moving `gradian/module` to its own package `common/` so it is
easily accessible to both card modules and the engine).