Improve client-backend communication
Previously, client-backend communication was very inconsistent
and lacking: very little information was communicated, and there
was not much consensus on what qualified as "significant
enough" to be communicated.

This commit improves client-backend communication by increasing
the number of events which are forwarded to the clients, and
increasing the information in pre-existing events.

The list of new/modified events include:

 * A player joining or leaving.
 * A status change (e.g., closing the game to new players).
 * The round starting.
 * The introductory message.