Link frontend and backend together
Link the frontend with the backend in a very messy manner. The
code is quite fragile and not well tested: due to the issues
with the backend API that I want to resolve, it seemed a bit
wasteful to spend a large portion of time ensuring that a UI
working with a poor API, only to have to redo all the work when
the API is refactored.